requirements.txt
# ElLocoGIS - AI-Driven Remote Sensing Semantic Search Engine
#
# Requirements for the application
# Install with: pip install -r requirements.txt

# Core Framework
streamlit>=1.28.0

# Geospatial & Earth Engine
geemap>=0.30.0
earthengine-api>=0.1.380
rasterio>=1.3.0
geopandas>=0.14.0
folium>=0.15.0
streamlit-folium>=0.15.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchgeo>=0.5.0
transformers>=4.35.0
open_clip_torch>=2.20.0
timm>=0.9.0
huggingface_hub>=0.19.0

# Vector Search
faiss-cpu>=1.7.4

# Data Processing
numpy>=1.24.0
pillow>=10.0.0
matplotlib>=3.7.0

# Utilities
requests>=2.31.0
tqdm>=4.66.0
